fix codecheck

+++ plugins/freeze_detector/resolver.cpp
@@ -57,14 +57,16 @@ bool FreezeResolver::ResolveEvent(const WatchPoint& watchPoint,
             list.push_back(watchPoint);
         } else if (window > 0) {
             unsigned long long start = timestamp;
-            unsigned long long end = timestamp + (window * MILLISECOND);
+            unsigned long long end = timestamp +
+                static_cast<unsigned long long>(window * MILLISECOND);
             std::string packageName = watchPoint.GetPackageName().empty() ?
                 watchPoint.GetProcessName() : watchPoint.GetPackageName();
             if (dBHelper_ != nullptr) {
                 dBHelper_->SelectEventFromDB(start, end, list, packageName, i);
             }
         } else {
-            unsigned long long start = timestamp + (window * MILLISECOND);
+            unsigned long long start = timestamp +
+                static_cast<unsigned long long>(window * MILLISECOND);
             unsigned long long end = timestamp;
             std::string packageName = watchPoint.GetPackageName().empty() ?
                 watchPoint.GetProcessName() : watchPoint.GetPackageName();
